rename resolution to magnification in shader relevant code
