+++ frontend/javascripts/oxalis/shaders/texture_access.glsl.ts
@@ -1,5 +1,5 @@
 import { MAX_ZOOM_STEP_DIFF } from "oxalis/model/bucket_data_handling/loading_strategy_logic";
-import { getResolutionFactors, getAbsoluteCoords } from "oxalis/shaders/coords.glsl";
+import { getMagnificationFactors, getAbsoluteCoords } from "oxalis/shaders/coords.glsl";
 import { hashCombine } from "./hashing.glsl";
 import type { ShaderModule } from "./shader_module_system";
 import { transDim } from "./utils.glsl";
@@ -92,7 +92,7 @@ export const getColorForCoords: ShaderModule = {
     getRgbaAtIndex,
     getRgbaAtXYIndex,
     getAbsoluteCoords,
-    getResolutionFactors,
+    getMagnificationFactors,
     hashCombine,
     transDim,
   ],
@@ -272,10 +272,10 @@ export const getColorForCoords: ShaderModule = {
          * If we are in the [5, _, _, 0]-bucket, we have to look into the **second** half
          * of the [2, _, _, 1]-bucket.
          * We can determine which "half" (subVolumeIndex) is relevant by doing a modulo operation
-         * with the resolution factor. A typical resolution factor is 2.
+         * with the magnification factor. A typical magnification factor is 2.
          */
 
-        vec3 magnificationFactors = getResolutionFactors(renderedMagIdx, activeMagIdx, globalLayerIndex);
+        vec3 magnificationFactors = getMagnificationFactors(renderedMagIdx, activeMagIdx, globalLayerIndex);
         vec3 coords = floor(getAbsoluteCoords(worldPositionUVW, activeMagIdx, globalLayerIndex));
         offsetInBucket = mod(coords, bucketWidth);
         vec3 worldBucketPosition = div(coords, bucketWidth);
