+++ frontend/javascripts/test/fixtures/mags.ts
