+++ frontend/javascripts/test/shaders/shader_syntax.spec.ts
@@ -1,6 +1,6 @@
 import "test/mocks/lz4";
 import getMainFragmentShader from "oxalis/shaders/main_data_shaders.glsl";
-import resolutions from "test/fixtures/resolutions";
+import mags from "test/fixtures/mags";
 import test, { type ExecutionContext } from "ava";
 import { parser } from "@shaderfrog/glsl-parser";
 
@@ -27,7 +27,7 @@ test("Shader syntax: Ortho Mode", (t: ExecutionContext<any>) => {
     },
     orderedColorLayerNames: ["color_layer_1", "color_layer_2"],
     segmentationLayerNames: [],
-    resolutionsCount: resolutions.length,
+    magnificationsCount: mags.length,
     voxelSizeFactor: [1, 1, 1],
     isOrthogonal: true,
     tpsTransformPerLayer: {},
@@ -54,7 +54,7 @@ test("Shader syntax: Ortho Mode + Segmentation - Mapping", (t: ExecutionContext<
     },
     orderedColorLayerNames: ["color_layer_1", "color_layer_2"],
     segmentationLayerNames: ["segmentationLayer"],
-    resolutionsCount: resolutions.length,
+    magnificationsCount: mags.length,
     voxelSizeFactor: [1, 1, 1],
     isOrthogonal: true,
     tpsTransformPerLayer: {},
@@ -74,7 +74,7 @@ test("Shader syntax: Ortho Mode + Segmentation + Mapping", (t: ExecutionContext<
     },
     orderedColorLayerNames: ["color_layer_1", "color_layer_2"],
     segmentationLayerNames: ["segmentationLayer"],
-    resolutionsCount: resolutions.length,
+    magnificationsCount: mags.length,
     voxelSizeFactor: [1, 1, 1],
     isOrthogonal: true,
     tpsTransformPerLayer: {},
@@ -94,7 +94,7 @@ test("Shader syntax: Arbitrary Mode (no segmentation available)", (t: ExecutionC
     },
     orderedColorLayerNames: ["color_layer_1", "color_layer_2"],
     segmentationLayerNames: [],
-    resolutionsCount: resolutions.length,
+    magnificationsCount: mags.length,
     voxelSizeFactor: [1, 1, 1],
     isOrthogonal: false,
     tpsTransformPerLayer: {},
@@ -114,7 +114,7 @@ test("Shader syntax: Arbitrary Mode (segmentation available)", (t: ExecutionCont
     },
     orderedColorLayerNames: ["color_layer_1", "color_layer_2"],
     segmentationLayerNames: ["segmentationLayer"],
-    resolutionsCount: resolutions.length,
+    magnificationsCount: mags.length,
     voxelSizeFactor: [1, 1, 1],
     isOrthogonal: false,
     tpsTransformPerLayer: {},
@@ -133,7 +133,7 @@ test("Shader syntax: Ortho Mode (rgb and float layer)", (t: ExecutionContext<any
     },
     orderedColorLayerNames: ["color_layer_1", "color_layer_2"],
     segmentationLayerNames: [],
-    resolutionsCount: resolutions.length,
+    magnificationsCount: mags.length,
     voxelSizeFactor: [1, 1, 1],
     isOrthogonal: true,
     tpsTransformPerLayer: {},
